Add early return

+++ lib/liquid/standardfilters.rb
@@ -143,7 +143,10 @@ module Liquid
     def url_remove_param(input, key)
       begin
         uri = URI.parse(input)
-        query = URI.decode_www_form(uri.query || '').to_h
+        
+        return uri.to_s if uri.query.empty?
+        
+        query = URI.decode_www_form(uri.query).to_h
         query.delete(key)
         uri.query = if query.empty?
           ''
